requirements.txt
discord.py==2.0.1
matplotlib==3.8.0
requests==2.31.0
